np.uint8 -> int

+++ stylized_facts/stylized_facts_checker.py
@@ -222,7 +222,7 @@ class StylizedFactsChecker:
         sampled_column: str = self.prng.choice(cumsum_scaled_transactions_df.columns)
         cumsum_scaled_transactions_arr: ndarray = cumsum_scaled_transactions_df[sampled_column].values
         cumsum_transactions_arr: ndarray = len(df) * cumsum_scaled_transactions_arr
-        cumsum_transactions: list[int] = list(cumsum_transactions_arr.astype(np.uint8))
+        cumsum_transactions: list[int] = list(cumsum_transactions_arr.astype(int))
         opens: list[Optional[float | int]] = []
         highes: list[Optional[float | int]] = []
         lowes: list[Optional[float | int]] = []
